refactor: extract theme editor hooks
